Split UringHandler off EpollUringHandler
