+++ io-uring-epoll/src/completion.rs
@@ -1,0 +1,27 @@
+//! Competion types
+
+use crate::slab::AcceptRec;
+use crate::slab::EpollRec;
+
+/// Completion types                      
+#[derive(Debug)]
+pub enum Completion {
+    /// EpollCtl Completion               
+    EpollEvent(EpollRec),
+    /// Accept Completion                 
+    Accept(AcceptRec),
+}
+
+/// What to do with the submission record upon handling completion.
+/// Used within handle_completions Fn Return                       
+#[derive(Clone, Debug, PartialEq)]
+pub enum SubmissionRecordStatus {
+    /// Retain the original submsision record when it is needed to be retained.
+    /// For example EpollCtl original Userdata must be retained in multishot mode.
+    /// Downside is that care must be taken to clean up the associated sunmission record.
+    Retain,
+    /// Forget the associated submission record                                          
+    /// For example Accept original record can be deleted upon compleiton after read.    
+    /// Typically a new Accept submission is pushed without re-using any existing.       
+    Forget,
+}
