+++ io-uring-epoll/src/error.rs
@@ -7,17 +7,26 @@ pub enum EpollHandlerError {
     Wait(String),
 }
 
-/// Errors from the handler           
+/// Errors from the Epoll Uring Handler
 #[derive(Debug)]
 pub enum EpollUringHandlerError {
-    /// Error creating IoUring instance
-    IoUringCreate(String),
-    /// Error creating epoll handle in Kernel
-    EpollCreate1(String),
+    /// Slab was able to store the value but not get it? This is a bug.
+    SlabBugSetGet(&'static str),
     /// EpollCtl OpCode is not supported in your kernel
     NotSupported,
     /// Error probing the support of EpollCtl from kernel
     Probing(String),
+    /// Error from the underlying UringHandler
+    UringHandler(UringHandlerError),
+    /// Error creating epoll handle in Kernel
+    EpollCreate1(String),
+}
+
+/// Errors from the Uring Handler
+#[derive(Debug)]
+pub enum UringHandlerError {
+    /// Error creating IoUring instance
+    IoUringCreate(String),
     /// The Fd is already in the handler and would override existing
     Duplicate,
     /// Something went yoinks in io_uring::IoRing::submit[_and_wait]
@@ -43,8 +52,7 @@ impl Display for EpollHandlerError {
 impl Display for EpollUringHandlerError {
     fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), std::fmt::Error> {
         match self {
-            Self::IoUringCreate(s) => write!(f, "IoUring Create: {}", s),
-            Self::EpollCreate1(s) => write!(f, "epoll_create1(): {}", s),
+            Self::SlabBugSetGet(s) => write!(f, "Slab Bug: {}", s),
             Self::NotSupported => write!(
                 f,
                 "EpollCtl io_uring OpCode is not supported in your Kernel"
@@ -54,6 +62,16 @@ impl Display for EpollUringHandlerError {
                 "Error whilst probing EpollCtl support from kernel: {}",
                 s
             ),
+            Self::UringHandler(s) => write!(f, "Underlying Uring error: {}", s),
+            Self::EpollCreate1(s) => write!(f, "epoll_create1(): {}", s),
+        }
+    }
+}
+
+impl Display for UringHandlerError {
+    fn fmt(&self, f: &mut Formatter<'_>) -> Result<(), std::fmt::Error> {
+        match self {
+            Self::IoUringCreate(s) => write!(f, "IoUring Create: {}", s),
             Self::Duplicate => write!(f, "The filehandle is already maped in. Possible duplicate?"),
             Self::Submission(s) => write!(f, "Submission: {}", s),
             Self::Slab(s) => write!(f, "Slab: {}", s),
